+++ atariari/methods/dim_baseline.py
@@ -66,7 +66,7 @@ class DIMTrainer(Trainer):
 
                 ts.append([t])
                 #thats.append([t_hat])
-            yield torch.stack(x_t).to(self.device) / 255., torch.stack(x_tprev).to(self.device) / 255.
+            yield torch.stack(x_t).float().to(self.device) / 255., torch.stack(x_tprev).float().to(self.device) / 255.
 
     def do_one_epoch(self, epoch, episodes):
         mode = "train" if self.encoder.training else "val"
