+++ atariari/methods/stdim.py
@@ -58,7 +58,7 @@ class InfoNCESpatioTemporalTrainer(Trainer):
 
                 x_tprev.append(episode[t - 1])
                 ts.append([t])
-            yield torch.stack(x_t).to(self.device) / 255., torch.stack(x_tprev).to(self.device) / 255.
+            yield torch.stack(x_t).float().to(self.device) / 255., torch.stack(x_tprev).float().to(self.device) / 255.
 
     def do_one_epoch(self, epoch, episodes):
         mode = "train" if self.encoder.training and self.classifier1.training else "val"
